requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-bootstrap-form==3.4
djangorestframework==3.13.1
flake8==4.0.1
idna==3.3
mccabe==0.6.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==5.3.0
